+++ test/integration/helpers.py
@@ -26,6 +26,7 @@ from os import environ, linesep, path
 from pathlib import Path
 from tempfile import gettempdir, mkdtemp
 from typing import List, Optional, Union
+from unittest.mock import MagicMock
 
 import backoff
 from b2sdk._v3.exception import BucketIdNotFound as v3BucketIdNotFound
@@ -248,8 +249,10 @@ def print_text_indented(text):
     """
     Prints text that may include weird characters, indented four spaces.
     """
+    mock_console_tool = MagicMock()
+    cmd_instance = Command(console_tool=mock_console_tool)
     for line in text.split(linesep):
-        Command._print_standard_descriptor(sys.stdout, '   ', repr(line)[1:-1])
+        cmd_instance._print_standard_descriptor(sys.stdout, '   ', repr(line)[1:-1])
 
 
 def print_output(status, stdout, stderr):